+++ src/configChangeListener.ts
@@ -1,7 +1,8 @@
 import * as vscode from "vscode";
 import { ConanAPI, ConanExecutionMode } from "./api/conan/conanAPI";
+import { ConfigurationManager } from "./configManager";
 
-export function configChangeListener(event: vscode.ConfigurationChangeEvent, conanApi: ConanAPI) {
+export function configChangeListener(event: vscode.ConfigurationChangeEvent, conanApi: ConanAPI, configManager: ConfigurationManager) {
     // TODO: Check if only vsconan configuration is changed, otherwise do nothing. Removing overhead.
 
     let pythonInterpreter: string = vscode.workspace.getConfiguration("vsconan").get("general.pythonInterpreter")!;
@@ -19,4 +20,24 @@ export function configChangeListener(event: vscode.ConfigurationChangeEvent, con
     else {
         conanApi.switchExecutionMode(ConanExecutionMode.conan);
     }
+
+    let conanUserHome: string = vscode.workspace.getConfiguration("vsconan").get("general.conanUserHome")!;
+
+    if (conanUserHome === null) { // Default value of configuration, see 'package.json'. Null means follow the pre defined environment variable
+        // Reset the current conan user home to the default environment variable
+        // We will get this from the config manager.
+        let envConanUserHome = configManager.getEnvConanUserHome();
+
+        if (envConanUserHome === undefined) {
+            // Deleting the environment variable, because it was not pre defined before
+            delete process.env.CONAN_USER_HOME;
+        }
+        else {
+            process.env.CONAN_USER_HOME = envConanUserHome;
+        }
+    }
+    else {
+        process.env.CONAN_USER_HOME = conanUserHome;
+    }
+
 }